backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.23.4
qrcode[pil]==7.4.2
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.1.0
requests==2.31.0
